Graduation rehearsal: ToggleButton moves labs → react intact
One intent file (core: minor, labs: major) covers both trains — the version
math maps the pre-1.0 labs major to a minor (fixture-tested). The component
folder moved with anatomy intact: pattern tests join the react suite
(9 passing), committed baselines verify unchanged at their new paths in the
pinned container (9 VRT+axe passing), manifests regenerate (react gains
toggle-button with experimental: false, labs empties), and labs' exports
drop the departed subpath. Reverted next — the rehearsal lives in history.
