+++ labs/src/index.ts
@@ -1,1 +1,2 @@
-export { ToggleButton, type ToggleButtonProps } from "./toggle-button/index.js";
+// Emptied by the toggle-button graduation rehearsal — the next experiment exports here.
+export {};
